change entity structure

+++ src/main/java/com/cloudcomputing/cloudcomputing/order/OrderService.java
@@ -2,13 +2,15 @@ package com.cloudcomputing.cloudcomputing.order;
 
 import com.cloudcomputing.cloudcomputing.ExceptionHandler.NotFoundException;
 import com.cloudcomputing.cloudcomputing.ExceptionHandler.ProductStockException;
+import com.cloudcomputing.cloudcomputing.ExceptionHandler.UnAuthorizedException;
+import com.cloudcomputing.cloudcomputing.business.BusinessMapper;
 import com.cloudcomputing.cloudcomputing.order.DTO.OrderDetailRequestDTO;
 import com.cloudcomputing.cloudcomputing.order.DTO.OrderRequest;
 import com.cloudcomputing.cloudcomputing.orderDetail.OrderDetail;
 import com.cloudcomputing.cloudcomputing.product.Product;
 import com.cloudcomputing.cloudcomputing.product.ProductRepository;
-import com.cloudcomputing.cloudcomputing.user.User;
-import com.cloudcomputing.cloudcomputing.user.UserRepository;
+import com.cloudcomputing.cloudcomputing.business.Business;
+import com.cloudcomputing.cloudcomputing.business.BusinessRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -24,14 +26,14 @@ import java.util.Optional;
 public class OrderService {
     private final OrderRepository orderRepository;
     private final ProductRepository productRepository;
-    private final UserRepository userRepository;
+    private final BusinessRepository businessRepository;
     @Autowired
     public OrderService(OrderRepository orderRepository,
                         ProductRepository productRepository,
-                        UserRepository userRepository) {
+                        BusinessRepository businessRepository) {
         this.orderRepository = orderRepository;
         this.productRepository = productRepository;
-        this.userRepository = userRepository;
+        this.businessRepository = businessRepository;
     }
 
     public Order findOneById (Long orderId){
@@ -46,22 +48,22 @@ public class OrderService {
         return orderRepository.findAllByOrderByOrderDateDesc(pageable);
     }
 
-    public List<Order> findAllByUser(Long userId){
-        return orderRepository.findByUserIdOrderByOrderDateDesc(userId);
+    public List<Order> findAllByUser(Long businessId){
+        return orderRepository.findByBusinessIdOrderByOrderDateDesc(businessId);
     }
 
 
 
 
-    public Order createOrder(OrderRequest request, Long userId){
+    public Order createOrder(OrderRequest request){
 
         // check user if it is present
-        Optional<User> user = userRepository.findById(userId);
-        if (user.isEmpty())
-            throw new NotFoundException("User with id: " +userId+" was not found");
+        Optional<Business> business = businessRepository.findById(request.getBusinessId());
+        if (business.isEmpty())
+            throw new NotFoundException("Business with id: " +request.getBusinessId()+" was not found");
 
         Order newOrder = new Order();
-        newOrder.setUser(user.get());
+        newOrder.setBusiness(business.get());
         List<OrderDetail> details = new ArrayList<>();
 
         List<Product> products = new ArrayList<>();
@@ -75,6 +77,9 @@ public class OrderService {
             //check product if it is present
             if (product.isEmpty())
                 throw new NotFoundException("Product with id: " +item.productId()+ " is not found");
+            if (!product.get().getBusiness().getName().equals(business.get().getName())){
+                throw new NotFoundException("Business: " +business.get().getName()+ " have no product with name: " + product.get().getName());            }
+
             products.add(product.get());
 
             totalCost += product.get().getPrice() * item.quantity();
@@ -95,7 +100,6 @@ public class OrderService {
         newOrder.setTotalCost(totalCost);
         newOrder.setOrderDetails(details);
 
-        user.get().getOrders().add(newOrder);
 
         products.forEach(product -> {
             int quantityOfItem = 0;
@@ -118,14 +122,19 @@ public class OrderService {
 
 
 
-    public void deleteOrderForBusiness(Long orderId){
+    public void deleteOrderForBusiness(Long orderId, Long businessId){
+        Optional<Business> business = businessRepository.findById(businessId);
+        if (business.isEmpty()) throw new NotFoundException("Business with id: " +businessId+ " is not found");
         Optional<Order> order = orderRepository.findById(orderId);
         if (order.isEmpty())
             throw new NotFoundException("Order with id: " +orderId+ " is not found");
+        if (!order.get().getBusiness().getName().equals(business.get().getName())){
+            throw new UnAuthorizedException("You are not permitted to delete this order");
+        }
         try{
             orderRepository.deleteById(orderId);
         }catch (Exception e){
-            throw new NotFoundException("Order with id: " +orderId+ " is not found");
+            throw new NotFoundException("Cant delete this order: " + e);
         }
     }
